+++ integreat_cms/cms/views/pages/__init__.py
@@ -23,3 +23,4 @@ from .page_actions import (
 from .page_sbs_view import PageSideBySideView
 from .page_revision_view import PageRevisionView
 from .page_xliff_import_view import PageXliffImportView
+from .partial_page_tree_view import PartialPageTreeView